# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebTui Chat Backend API
  version: 0.9.0
  description: Contract nền cho backend WebTui Chat. Tài liệu API được cập nhật theo từng phase backend.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{workspace_id}:
    get:
      summary: Xem chi tiết workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Chi tiết workspace
    patch:
      summary: Cập nhật workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace đã cập nhật
    delete:
      summary: Xóa mềm workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "204":
          description: Đã xóa mềm
...
  /api/v1/workspaces/{workspace_id}/files:
    get:
      summary: Danh sách file trong workspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Danh sách file
    post:
      summary: Upload file bằng multipart form
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                metadata:
                  type: string
                  description: JSON string tùy chọn.
      responses:
        "201":
          description: Upload file thành công
  /api/v1/workspaces/{workspace_id}/files/{file_id}:
    get:
      summary: Xem metadata file
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: Metadata file
  /api/v1/workspaces/{workspace_id}/files/{file_id}/download:
    get:
      summary: Tải nội dung file
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: Nội dung file
  /api/v1/workspaces/{workspace_id}/files/{file_id}/versions:
    get:
      summary: Danh sách version của file
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          description: Danh sách version
    post:
      summary: Upload version mới cho file
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/FileID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Tạo version mới thành công
...
  /api/v1/workspaces/{workspace_id}/channels/{channel_id}/messages/{message_id}/attachments:
    get:
      summary: Danh sách file gắn với tin nhắn
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/ChannelID"
        - $ref: "#/components/parameters/MessageID"
      responses:
        "200":
          description: Danh sách attachment
    post:
      summary: Gắn file vào tin nhắn
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/ChannelID"
        - $ref: "#/components/parameters/MessageID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_id]
              properties:
                file_id:
                  type: string
                  format: uuid
                sort_order:
                  type: integer
      responses:
        "201":
          description: File đã được gắn vào tin nhắn
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChannelID:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MessageID:
      name: message_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileID:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...